Bash script version of what I'm trying to do
It works! (Debugging the off-by-one in the header segment took me
forever...)

But this requires the user to download and/or install both of:
1. https://github.com/AppleCommander/bastools/releases/latest
2. https://github.com/datajerk/c2t/blob/master/bin/c2t

It'd be nice to hide the entire .bas program to .wav file conversion
behind a single command, without any dependencies.

I think version 1 of the workflow could look like this:
```
$ git clone https://github.com/khollbach/bas2wav
$ cargo run my-program.bas
```

And for version 2, figure out how to use cargo for package management:
```
$ cargo install bas2wav
$ bas2wav my-program.bas
```